requirements.txt
boto3==1.38.33
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-redis==5.4.0
django-storages==1.14.6
django_solvitize==0.0.6
djangorestframework==3.16.0
gunicorn==23.0.0
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.4
sentry-sdk==2.28.0
whitenoise

